pyproject.toml
[tool.poetry]
name = "hw-27"
version = "0.1.0"
description = ""
authors = ["SaixFix <maxim-grishns@mail.ru>"]
readme = "README.md"
packages = [{include = "hw_27"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
